сервисы и контроллеры разделены по правам доступа: public, private, admin
