+++ main-server/src/main/java/ru/practicum/ewm/comment/controller/CommentPublicController.java
@@ -1,0 +1,26 @@
+package ru.practicum.ewm.comment.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+import ru.practicum.ewm.comment.dto.CommentDto;
+import ru.practicum.ewm.comment.service.CommentPublicService;
+
+import javax.validation.constraints.Positive;
+import javax.validation.constraints.PositiveOrZero;
+import java.util.Collection;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping
+public class CommentPublicController {
+    private final CommentPublicService service;
+
+    @GetMapping("/events/{eventId}/comments")
+    public Collection<CommentDto> findEventComments(@PathVariable long eventId,
+                                                    @RequestParam(required = false) String rangeStart,
+                                                    @RequestParam(required = false) String rangeEnd,
+                                                    @PositiveOrZero @RequestParam(defaultValue = "0") int from,
+                                                    @Positive @RequestParam(defaultValue = "10") int size) {
+        return service.getPublishedEventComments(eventId, rangeStart, rangeEnd, from, size);
+    }
+}
