+++ main-server/src/main/java/ru/practicum/ewm/comment/dto/CommentDto.java
@@ -2,7 +2,7 @@ package ru.practicum.ewm.comment.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import ru.practicum.ewm.comment.CommentState;
+import ru.practicum.ewm.comment.model.CommentState;
 
 import java.time.LocalDateTime;
