+++ main-server/src/main/java/ru/practicum/ewm/comment/model/Comment.java
@@ -1,13 +1,17 @@
-package ru.practicum.ewm.comment;
+package ru.practicum.ewm.comment.model;
 
-import lombok.Data;
+import lombok.*;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
 
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity
 @Table(name = "comments", schema = "public")
-@Data
 public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
