+++ main-server/src/main/java/ru/practicum/ewm/comment/model/CommentState.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.comment;
+package ru.practicum.ewm.comment.model;
 
 public enum CommentState {
     PENDING, PUBLISHED, CANCELED
