+++ main-server/src/main/java/ru/practicum/ewm/comment/repository/CommentRepository.java
@@ -1,10 +1,11 @@
-package ru.practicum.ewm.comment;
+package ru.practicum.ewm.comment.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import ru.practicum.ewm.comment.model.Comment;
 
 import java.time.LocalDateTime;
 
@@ -17,7 +18,7 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
 
     @Query(value = " with recursive comms (id, path, level ) " +
             "as ( select c.id, " +
-            "            to_char(c.published, 'yyyy-MM-dd HH:mm:ss') ||'->'|| cast (c.id as varchar(50)) as path, 1 " +
+            "            to_char(c.published, 'yyyy-mm-dd HH24:mi:ss') ||'->'|| cast (c.id as varchar(50)) as path, 1 " +
             "       from comments c " +
             "      where c.event_id = :eventId " +
             "        and c.published between :rangeStart and :rangeEnd " +
@@ -26,7 +27,7 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
             "      union all " +
             "     select c.id, cast (comms.path ||'->'|| c.id as varchar(50)) as path, level + 1 " +
             "       from comments c inner join comms on (c.comment_id = comms.id and c.state = 'PUBLISHED')) " +
-            "select c.* from comments c inner join comms on (c.id = comms.id) order by comms.path limit 1000 ",
+            "select c.* from comments c inner join comms on (c.id = comms.id) order by comms.path ",
             nativeQuery = true)
     Page<Comment> findPublishedEventComments(@Param("eventId") Long eventId,
                                              @Param("rangeStart") LocalDateTime rangeStart,
