+++ main-server/src/main/java/ru/practicum/ewm/comment/service/CommentPublicService.java
@@ -1,0 +1,13 @@
+package ru.practicum.ewm.comment.service;
+
+import ru.practicum.ewm.comment.dto.CommentDto;
+
+import java.util.Collection;
+
+public interface CommentPublicService {
+    Collection<CommentDto> getPublishedEventComments(Long eventId,
+                                                     String rangeStart,
+                                                     String rangeEnd,
+                                                     int from,
+                                                     int size);
+}
